+++ crates/nostr-relay-builder/src/builder.rs
@@ -35,10 +35,18 @@ impl Default for RateLimit {
     }
 }
 
+#[cfg(feature = "tor")]
+#[allow(missing_docs)]
+#[deprecated(
+    since = "0.45.0",
+    note = "Use `LocalRelayBuilderHiddenService` instead"
+)]
+pub type RelayBuilderHiddenService = LocalRelayBuilderHiddenService;
+
 /// Relay builder tor hidden service options
 #[derive(Debug, Clone)]
 #[cfg(feature = "tor")]
-pub struct RelayBuilderHiddenService {
+pub struct LocalRelayBuilderHiddenService {
     /// Nickname (local identifier) for a Tor hidden service
     ///
     /// Used to look up this service's keys, state, configuration, etc., and distinguish them from other services.
@@ -48,7 +56,7 @@ pub struct RelayBuilderHiddenService {
 }
 
 #[cfg(feature = "tor")]
-impl RelayBuilderHiddenService {
+impl LocalRelayBuilderHiddenService {
     /// New tor hidden service options
     ///
     /// The nickname is a local identifier for a Tor hidden service.
@@ -83,9 +91,13 @@ impl RelayBuilderHiddenService {
     }
 }
 
+#[allow(missing_docs)]
+#[deprecated(since = "0.45.0", note = "Use `LocalRelayBuilderMode` instead")]
+pub type RelayBuilderMode = LocalRelayBuilderMode;
+
 /// Mode
 #[derive(Debug, Clone, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum RelayBuilderMode {
+pub enum LocalRelayBuilderMode {
     /// Generic mode
     #[default]
     Generic,
@@ -123,18 +135,26 @@ pub trait QueryPolicy: fmt::Debug + Send + Sync {
     ) -> BoxedFuture<'a, PolicyResult>;
 }
 
+#[allow(missing_docs)]
+#[deprecated(since = "0.45.0", note = "Use `LocalRelayTestOptions` instead")]
+pub type RelayTestOptions = LocalRelayTestOptions;
+
 /// Testing options
 #[derive(Debug, Clone, Default)]
-pub struct RelayTestOptions {
+pub struct LocalRelayTestOptions {
     /// Simulate unresponsive connection
     pub unresponsive_connection: Option<Duration>,
     /// Send random events to the clients
     pub send_random_events: bool,
 }
 
+#[allow(missing_docs)]
+#[deprecated(since = "0.45.0", note = "Use `LocalRelayBuilderNip42Mode` instead")]
+pub type RelayBuilderNip42Mode = LocalRelayBuilderNip42Mode;
+
 /// NIP42 mode
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum RelayBuilderNip42Mode {
+pub enum LocalRelayBuilderNip42Mode {
     /// Require authentication for writing
     Write,
     /// Require authentication for reading
@@ -144,32 +164,40 @@ pub enum RelayBuilderNip42Mode {
     Both,
 }
 
-impl RelayBuilderNip42Mode {
-    /// Check if is [`RelayBuilderNip42Mode::Read`] or [`RelayBuilderNip42Mode::Both`]
+impl LocalRelayBuilderNip42Mode {
+    /// Check if is [`LocalRelayBuilderNip42Mode::Read`] or [`LocalRelayBuilderNip42Mode::Both`]
     #[inline]
     pub fn is_read(&self) -> bool {
         matches!(self, Self::Read | Self::Both)
     }
 
-    /// Check if is [`RelayBuilderNip42Mode::Write`] or [`RelayBuilderNip42Mode::Both`]
+    /// Check if is [`LocalRelayBuilderNip42Mode::Write`] or [`LocalRelayBuilderNip42Mode::Both`]
     #[inline]
     pub fn is_write(&self) -> bool {
         matches!(self, Self::Write | Self::Both)
     }
 }
 
+#[allow(missing_docs)]
+#[deprecated(since = "0.45.0", note = "Use `LocalRelayBuilderNip42` instead")]
+pub type RelayBuilderNip42 = LocalRelayBuilderNip42;
+
 /// NIP42 options
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct RelayBuilderNip42 {
+pub struct LocalRelayBuilderNip42 {
     /// Mode
-    pub mode: RelayBuilderNip42Mode,
+    pub mode: LocalRelayBuilderNip42Mode,
     // /// Allowed public keys
     // pub allowed: HashSet<PublicKey>,
 }
 
-/// Relay builder
+#[allow(missing_docs)]
+#[deprecated(since = "0.45.0", note = "Use `LocalRelayBuilder` instead")]
+pub type RelayBuilder = LocalRelayBuilder;
+
+/// Local relay builder
 #[derive(Debug, Clone)]
-pub struct RelayBuilder {
+pub struct LocalRelayBuilder {
     /// IP address
     pub(crate) addr: Option<IpAddr>,
     /// Port
@@ -177,14 +205,14 @@ pub struct RelayBuilder {
     /// Database
     pub(crate) database: Arc<dyn NostrDatabase>,
     /// Mode
-    pub(crate) mode: RelayBuilderMode,
+    pub(crate) mode: LocalRelayBuilderMode,
     /// Rate limit
     pub(crate) rate_limit: RateLimit,
     /// NIP42 options
-    pub(crate) nip42: Option<RelayBuilderNip42>,
+    pub(crate) nip42: Option<LocalRelayBuilderNip42>,
     /// Tor hidden service
     #[cfg(feature = "tor")]
-    pub(crate) tor: Option<RelayBuilderHiddenService>,
+    pub(crate) tor: Option<LocalRelayBuilderHiddenService>,
     /// Max connections allowed
     pub(crate) max_connections: Option<usize>,
     /// Max subscription ID length
@@ -203,10 +231,10 @@ pub struct RelayBuilder {
     /// Query policy plugins
     pub(crate) query_plugins: Vec<Arc<dyn QueryPolicy>>,
     /// Test options
-    pub(crate) test: RelayTestOptions,
+    pub(crate) test: LocalRelayTestOptions,
 }
 
-impl Default for RelayBuilder {
+impl Default for LocalRelayBuilder {
     fn default() -> Self {
         Self {
             addr: None,
@@ -215,7 +243,7 @@ impl Default for RelayBuilder {
                 events: true,
                 max_events: Some(NonZeroUsize::new(75_000).unwrap()),
             })),
-            mode: RelayBuilderMode::default(),
+            mode: LocalRelayBuilderMode::default(),
             rate_limit: RateLimit::default(),
             nip42: None,
             #[cfg(feature = "tor")]
@@ -228,12 +256,12 @@ impl Default for RelayBuilder {
             min_pow: None,
             write_plugins: Vec::new(),
             query_plugins: Vec::new(),
-            test: RelayTestOptions::default(),
+            test: LocalRelayTestOptions::default(),
         }
     }
 }
 
-impl RelayBuilder {
+impl LocalRelayBuilder {
     /// Set IP address
     #[inline]
     pub fn addr(mut self, ip: IpAddr) -> Self {
@@ -260,7 +288,7 @@ impl RelayBuilder {
 
     /// Set mode
     #[inline]
-    pub fn mode(mut self, mode: RelayBuilderMode) -> Self {
+    pub fn mode(mut self, mode: LocalRelayBuilderMode) -> Self {
         self.mode = mode;
         self
     }
@@ -274,7 +302,7 @@ impl RelayBuilder {
 
     /// Require NIP42 authentication
     #[inline]
-    pub fn nip42(mut self, opts: RelayBuilderNip42) -> Self {
+    pub fn nip42(mut self, opts: LocalRelayBuilderNip42) -> Self {
         self.nip42 = Some(opts);
         self
     }
@@ -282,7 +310,7 @@ impl RelayBuilder {
     /// Set tor options
     #[inline]
     #[cfg(feature = "tor")]
-    pub fn tor(mut self, opts: RelayBuilderHiddenService) -> Self {
+    pub fn tor(mut self, opts: LocalRelayBuilderHiddenService) -> Self {
         self.tor = Some(opts);
         self
     }
@@ -357,7 +385,7 @@ impl RelayBuilder {
 
     /// Testing options
     #[inline]
-    pub(crate) fn test(mut self, test: RelayTestOptions) -> Self {
+    pub(crate) fn test(mut self, test: LocalRelayTestOptions) -> Self {
         self.test = test;
         self
     }