+++ crates/nostr-relay-builder/src/lib.rs
@@ -16,7 +16,7 @@ pub mod local;
 pub mod mock;
 pub mod prelude;
 
-pub use self::builder::RelayBuilder;
+pub use self::builder::LocalRelayBuilder;
 pub use self::error::Error;
 pub use self::local::LocalRelay;
 pub use self::mock::MockRelay;