+++ crates/nostr-relay-builder/CHANGELOG.md
@@ -23,6 +23,12 @@
 
 -->
 
+## Unreleased
+
+### Changes
+
+- Rename all `RelayBuilder*` structs and enums to `LocalRelayBuilder*` (https://github.com/rust-nostr/nostr/pull/1145)
+
 ## v0.44.0 - 2025/11/06
 
 ### Breaking change